Transform to component then move to internal package
